Updated styling, app title
The auth page is now prettier. There were some extra quotation
marks and stuff that was messing up the style. So I overhauled
it. This also implements basic responsiveness to window width.
You can choose to view it in 3-column mode or 1-column mode
by changing your screen width, which is neat.