Added a property to test layouts, changed layout
I added the 'desc' property to the test array to play with putting
a snippet of the pull description in the layout. I didn't come up
with anything that I liked yet, but I did end up making the link
the pull # and adding a <p> tag to hold the pull name, as I noticed
that the full pull name would be quite long and fill up the column
pretty quickly.